refactor(web): consolidate backend into web service
