+++ .env.example
@@ -26,11 +26,9 @@ REALTIME_INTERNAL_TOKEN=replace-with-at-least-32-random-characters
 # Optional comma-separated browser origins allowed to open a socket.
 REALTIME_ALLOWED_ORIGINS=http://localhost:3000
 BETTER_AUTH_SECRET=replace-with-at-least-32-random-characters
-# Public API origin used by Better Auth when it creates callback URLs.
-BETTER_AUTH_URL=http://localhost:3001
-# Browser app origin and credentialed CORS allowlist for the standalone API.
+BETTER_AUTH_URL=http://localhost:3000
+# Public browser origin used by auth, billing, and external redirects.
 APP_ORIGIN=http://localhost:3000
-API_ALLOWED_ORIGINS=http://localhost:3000,https://loora.design
 REQUIRE_PREVIEW_ACCESS=true
 REQUIRE_POLAR_BILLING=false
 POLAR_SERVER=production
@@ -93,13 +91,13 @@ MCP_SCREENSHOT_QUEUE_TIMEOUT_MS=20000
 MCP_SCREENSHOT_IDLE_TIMEOUT_MS=60000
 MCP_AUTH_TIMEOUT_MS=5000
 MCP_INTERNAL_TOKEN=replace-with-a-distinct-at-least-32-character-secret
-# Optional private API-service URL; defaults to BETTER_AUTH_URL/api/internal/mcp.
+# Optional private web-service URL; defaults to BETTER_AUTH_URL/api/internal/mcp.
 MCP_INTERNAL_API_URL=
-# Opt-in JSON request timing logs for web/API/MCP (`api.request` console.info lines).
+# Opt-in JSON request timing logs for web/MCP (`api.request` console.info lines).
 LOG_REQUEST_TIMING=false
 # Local authenticated API benchmark. Keep credential values out of commits.
 LOORA_BENCHMARK_TARGET=canvas
-LOORA_BENCHMARK_BASE_URL=http://localhost:3001
+LOORA_BENCHMARK_BASE_URL=http://localhost:3000
 LOORA_BENCHMARK_DESIGN_ID=
 LOORA_BENCHMARK_DRAFT_ID=
 LOORA_BENCHMARK_SESSION_COOKIE=
