--- apps/api/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@loora/api",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "bun --env-file=../../.env --watch run src/index.ts",
-    "start": "bun --env-file=../../.env run src/index.ts",
-    "typecheck": "bunx tsc --noEmit -p tsconfig.json"
-  },
-  "dependencies": {
-    "@loora/auth": "workspace:*",
-    "@loora/billing": "workspace:*",
-    "@loora/db": "workspace:*",
-    "@loora/realtime": "workspace:*",
-    "@loora/rpc": "workspace:*",
-    "@orpc/server": "^1.14.8",
-    "drizzle-orm": "catalog:",
-    "hono": "^4.9.6"
-  },
-  "devDependencies": {
-    "@types/bun": "^1.3.14",
-    "typescript": "catalog:"
-  }
-}
