--- apps/api/railway.json
@@ -1,22 +1,0 @@
-{
-  "$schema": "https://railway.app/railway.schema.json",
-  "build": {
-    "builder": "DOCKERFILE",
-    "dockerfilePath": "apps/api/Dockerfile",
-    "watchPatterns": [
-      "apps/api/**",
-      "packages/**",
-      "package.json",
-      "bun.lock",
-      "bunfig.toml",
-      "tsconfig.json",
-      "apps/api/railway.json"
-    ]
-  },
-  "deploy": {
-    "restartPolicyType": "ON_FAILURE",
-    "restartPolicyMaxRetries": 5,
-    "sleepApplication": false,
-    "healthcheckPath": "/api/ready"
-  }
-}
