+++ package.json
@@ -14,7 +14,6 @@
   "private": true,
   "scripts": {
     "dev": "bun run --cwd apps/web dev",
-    "dev:api": "bun run --cwd apps/api dev",
     "dev:desktop": "bun run --cwd apps/desktop dev",
     "dev:ws": "cargo run -p loora-ws-server",
     "dev:mcp": "cargo run -p loora-mcp-server",
@@ -28,7 +27,6 @@
     "docs:build": "bun run --cwd apps/docs build",
     "preview": "bun run --cwd apps/web preview",
     "start": "bun run --cwd apps/web start",
-    "start:api": "bun run --cwd apps/api start",
     "test": "vitest run",
     "generate-routes": "bun run --cwd apps/web generate-routes",
     "db:generate": "bun run --cwd packages/db generate",
