+++ packages/rpc/src/client.ts
@@ -8,9 +8,9 @@ import type { appRouter } from './router.ts'
  * The browser half of the router. `appRouter` is a type-only import, so none
  * of the server implementation follows it into the client bundle.
  *
- * The URL comes from the platform runtime: the web app calls the configured
- * API origin, and the desktop app calls the loopback server in its own process,
- * which proxies on to Loora with the session it holds.
+ * The URL comes from the platform runtime: the web app calls its own origin,
+ * and the desktop app calls the loopback server in its own process, which
+ * proxies on to Loora with the session it holds.
  */
 const link = new RPCLink({
   url: () => apiUrl('/api/rpc'),
